Moved move_file funct to system, adds backwards compatibility to env change
